requirements_models.txt
# Additional requirements for model training

# Tinker SDK for LLM LoRA fine-tuning
tinker

# Transformers and PyTorch (kept for local/fallback BERT/RoBERTa path)
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
tokenizers>=0.13.0
accelerate>=0.18.0

# Class imbalance handling
imbalanced-learn>=0.11.0

# API integration
requests>=2.31.0
python-dotenv>=1.0.0

# Additional ML utilities
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization and monitoring
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Utilities
tqdm>=4.66.0
joblib>=1.3.0
